requirements.txt
PyQt6==6.7.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
Jinja2==3.1.4
WeasyPrint==61.2
APScheduler==3.10.4
Babel==2.15.0
alembic==1.13.2

python-escpos==2.2.0
bcrypt==4.2.0
Pillow==10.4.0

pyusb==1.2.1
matplotlib==3.9.2
